requirements.txt
uvicorn~=0.22.0
fastapi~=0.95.2
python-dotenv~=1.0.0
langchain~=0.0.174
pydantic~=1.10.7
azure-core~=1.26.4
tiktoken~=0.4.0
openai~=0.27.7
pdf2image~=1.16.3
unstructured
chromadb

